LPS-92556 Move to SPI
