+++ modules/apps/data-engine/data-engine-spi/src/main/resources/com/liferay/data/engine/spi/rule/function/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
